Moves files into a new physical structure to match solution Filters (modified in commit 986da76)
