+++ src/TimeTrace/ExecutionHierarchy.cpp
