+++ src/TimeTrace/TimeTraceGenerator.cpp
