+++ src/WPA/Analyzers/MiscellaneousCache.h
