+++ vcperf.vcxproj.filters
@@ -48,31 +48,31 @@
     <ClCompile Include="src\GenericFields.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="src\Views\BuildExplorerView.cpp">
+    <ClCompile Include="src\WPA\Views\BuildExplorerView.cpp">
       <Filter>Source Files\WPA\Views</Filter>
     </ClCompile>
-    <ClCompile Include="src\Views\FilesView.cpp">
+    <ClCompile Include="src\WPA\Views\FilesView.cpp">
       <Filter>Source Files\WPA\Views</Filter>
     </ClCompile>
-    <ClCompile Include="src\Views\FunctionsView.cpp">
+    <ClCompile Include="src\WPA\Views\FunctionsView.cpp">
       <Filter>Source Files\WPA\Views</Filter>
     </ClCompile>
-    <ClCompile Include="src\Views\TemplateInstantiationsView.cpp">
+    <ClCompile Include="src\WPA\Views\TemplateInstantiationsView.cpp">
       <Filter>Source Files\WPA\Views</Filter>
     </ClCompile>
-    <ClCompile Include="src\Analyzers\ContextBuilder.cpp">
+    <ClCompile Include="src\WPA\Analyzers\ContextBuilder.cpp">
       <Filter>Source Files\WPA\Analyzers</Filter>
     </ClCompile>
-    <ClCompile Include="src\Analyzers\ExpensiveTemplateInstantiationCache.cpp">
+    <ClCompile Include="src\WPA\Analyzers\ExpensiveTemplateInstantiationCache.cpp">
       <Filter>Source Files\WPA\Analyzers</Filter>
     </ClCompile>
-    <ClCompile Include="src\Views\TimeTrace\TimeTraceGenerator.cpp">
+    <ClCompile Include="src\TimeTrace\TimeTraceGenerator.cpp">
       <Filter>Source Files\TimeTrace</Filter>
     </ClCompile>
-    <ClCompile Include="src\Views\TimeTrace\PackedProcessThreadRemapping.cpp">
+    <ClCompile Include="src\TimeTrace\PackedProcessThreadRemapping.cpp">
       <Filter>Source Files\TimeTrace</Filter>
     </ClCompile>
-    <ClCompile Include="src\Analyzers\ExecutionHierarchy.cpp">
+    <ClCompile Include="src\TimeTrace\ExecutionHierarchy.cpp">
       <Filter>Source Files\TimeTrace</Filter>
     </ClCompile>
   </ItemGroup>
@@ -89,37 +89,37 @@
     <ClInclude Include="src\GenericFields.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="src\Views\BuildExplorerView.h">
+    <ClInclude Include="src\WPA\Views\BuildExplorerView.h">
       <Filter>Header Files\WPA\Views</Filter>
     </ClInclude>
-    <ClInclude Include="src\Views\FunctionsView.h">
+    <ClInclude Include="src\WPA\Views\FunctionsView.h">
       <Filter>Header Files\WPA\Views</Filter>
     </ClInclude>
-    <ClInclude Include="src\Views\TemplateInstantiationsView.h">
+    <ClInclude Include="src\WPA\Views\TemplateInstantiationsView.h">
       <Filter>Header Files\WPA\Views</Filter>
     </ClInclude>
-    <ClInclude Include="src\Views\FilesView.h">
+    <ClInclude Include="src\WPA\Views\FilesView.h">
       <Filter>Header Files\WPA\Views</Filter>
     </ClInclude>
-    <ClInclude Include="src\Analyzers\ExpensiveTemplateInstantiationCache.h">
+    <ClInclude Include="src\WPA\Analyzers\ExpensiveTemplateInstantiationCache.h">
       <Filter>Header Files\WPA\Analyzers</Filter>
     </ClInclude>
-    <ClInclude Include="src\Analyzers\MiscellaneousCache.h">
+    <ClInclude Include="src\WPA\Analyzers\MiscellaneousCache.h">
       <Filter>Header Files\WPA\Analyzers</Filter>
     </ClInclude>
-    <ClInclude Include="src\Analyzers\ContextBuilder.h">
+    <ClInclude Include="src\WPA\Analyzers\ContextBuilder.h">
       <Filter>Header Files\WPA\Analyzers</Filter>
     </ClInclude>
     <ClInclude Include="src\VcperfBuildInsights.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="src\Views\TimeTrace\PackedProcessThreadRemapping.h">
+    <ClInclude Include="src\TimeTrace\PackedProcessThreadRemapping.h">
       <Filter>Header Files\TimeTrace</Filter>
     </ClInclude>
-    <ClInclude Include="src\Views\TimeTrace\TimeTraceGenerator.h">
+    <ClInclude Include="src\TimeTrace\TimeTraceGenerator.h">
       <Filter>Header Files\TimeTrace</Filter>
     </ClInclude>
-    <ClInclude Include="src\Analyzers\ExecutionHierarchy.h">
+    <ClInclude Include="src\TimeTrace\ExecutionHierarchy.h">
       <Filter>Header Files\TimeTrace</Filter>
     </ClInclude>
   </ItemGroup>
